fix: restore Fragment-based architecture and fix compilation errors
- Restore original MainActivity structure with Fragment support
- Fix type mismatch in NativeEngine (channels.toShort())
- Add NativeAudioManager utility for Fragment-native engine interaction
- Keep native engine initialization without breaking existing UI
- Fragments can now access native engine via NativeAudioManager.from(fragment)

+++ app/src/main/java/com/georgv/audioworkstation/engine/NativeEngine.kt
@@ -44,7 +44,7 @@ class NativeEngine {
 				fos.write("fmt ".toByteArray())
 				writeInt32LE(fos, 16) // fmt chunk size
 				writeInt16LE(fos, 1) // audio format (PCM)
-				writeInt16LE(fos, channels) // num channels
+				writeInt16LE(fos, channels.toShort()) // num channels
 				writeInt32LE(fos, sampleRate) // sample rate
 				writeInt32LE(fos, sampleRate * channels * 2) // byte rate
 				writeInt16LE(fos, (channels * 2).toShort()) // block align
